fixed type conflict between timeValid paramenters and arguments - converted all parameters to date object for starTime

+++ timeblock.h
@@ -18,7 +18,7 @@ public:
     ~TimeBlock();
 private:
     Task* task;
-    float tbStartTime;
+    double tbStartTime;
     //float increment = 0.25;
 };
